requirements.txt
pandas
msal
requests
python-dotenv
openpyxl
pyyaml